backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
fabrictestbed-extensions @ git+https://github.com/fabric-testbed/fabrictestbed-extensions.git@fabvis
pydantic>=2.0
python-multipart>=0.0.9
websockets>=12.0
paramiko>=3.4
requests>=2.31
httpx>=0.27
openai>=1.0
duckduckgo-search>=7.0
jupyterlab
jupyter-ai
ipykernel
pandas
matplotlib
numpy
plotly
scipy
seaborn
ipywidgets
pytest-cov
pytest-asyncio
